cleanup the strings, add headers
