+++ src/NuGet.Clients/NuGet.PackageManagement.UI/NuGet.PackageManagement.UI.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), 'README.md'))\build\common.legacy.props" />
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
@@ -130,6 +130,7 @@
     </Compile>
     <Compile Include="Xamls\InfiniteScrollListBox.cs" />
     <Compile Include="Xamls\InfiniteScrollListBoxAutomationPeer.cs" />
+    <Compile Include="Models\LicenseFileData.cs" />
     <Compile Include="Xamls\LicenseFileWindow.xaml.cs">
       <DependentUpon>LicenseFileWindow.xaml</DependentUpon>
     </Compile>
