+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Xamls/LicenseFileWindow.xaml
@@ -17,12 +17,19 @@
   d:DesignHeight="300"
   d:DesignWidth="300"
   >
+  <Window.Resources>
+    <ResourceDictionary>
+      <ResourceDictionary.MergedDictionaries>
+        <nuget:SharedResources />
+      </ResourceDictionary.MergedDictionaries>
+    </ResourceDictionary>
+  </Window.Resources>
+
   <Grid>
     <Border
       BorderThickness="1"
       BorderBrush="{DynamicResource {x:Static nuget:Brushes.BorderBrush}}">
       <ScrollViewer
-              Padding="3"
               CanContentScroll="True"
               VerticalScrollBarVisibility="Visible"
               HorizontalScrollBarVisibility="Disabled">
