+++ src/NuGet.Core/NuGet.Protocol/Strings.resx
@@ -500,4 +500,15 @@ The "ms" should be localized to the abbreviation for milliseconds.</comment>
   <data name="NuGetLicenseExpression_NonStandardIdentifier" xml:space="preserve">
     <value>The license identifier(s) {0} is(are) not recognized by the current toolset.</value>
   </data>
+  <data name="LoadFileFromNupkg_FileNotFound" xml:space="preserve">
+    <value>The file '{0}' does not exist in the package.</value>
+    <comment>0 - file path relative to the nupkg.</comment>
+  </data>
+  <data name="LoadFileFromNupkg_FileTooLarge" xml:space="preserve">
+    <value>Cannot load file. The file is larger than 100MB.</value>
+  </data>
+  <data name="LoadFileFromNupkg_UnknownProblemLoadingTheFile" xml:space="preserve">
+    <value>Unknown problem loading the file '{0}'.</value>
+    <comment>0 - file path relative to the nupkg.</comment>
+  </data>
 </root>
